+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 [_vNext_](https://github.com/sketch7/SignalR.Orleans/compare/1.0.0...1.1.0) (2024-X-X)
 
+## [9.0.0](https://github.com/sketch7/SignalR.Orleans/compare/8.0.1...9.0.0) (2026-04-17)
+
+### Features
+
+- **deps:** update to Orleans 10.1.0
+
 ## [8.0.1](https://github.com/sketch7/SignalR.Orleans/compare/8.0.0...8.0.1) (2025-11-27)
 
 ### Features